src/main/java/org/leonidasanin/algorithmbasedsolvedtasks/controller/ExceptionHandlerController.java
package org.leonidasanin.algorithmbasedsolvedtasks.controller;

import org.leonidasanin.algorithmbasedsolvedtasks.exception.TaskException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.servlet.mvc.support.RedirectAttributes;

import java.io.IOException;

@ControllerAdvice
public class ExceptionHandlerController {

    @ExceptionHandler(TaskException.class)
    public String handleTaskException(TaskException exception, RedirectAttributes redirectAttributes) {
        redirectAttributes.addFlashAttribute("task", exception.getPreviousTask());
        redirectAttributes.addFlashAttribute("error", exception.getMessage());
        return "redirect:/";
    }

    @ExceptionHandler(IOException.class)
    public String handleIOException(IOException exception, RedirectAttributes redirectAttributes) {
        redirectAttributes.addFlashAttribute("error", exception.getMessage());
        return "redirect:/";
    }
}
